+++ all/file_exists/index.browser.js
@@ -1,4 +1,3 @@
-import { waitfor } from '../waitfor/index.browser.js'
 /**
  * @param {string}path
  * @returns {Promise<boolean>}
@@ -11,18 +10,15 @@ export {
 	file_exists_ as path__exists_
 }
 /**
- * @param {string}path
+ * @param {string|(()=>unknown|Promise<unknown>)}path_OR_op
  * @param {number}[timeout]
  * @param {number|((promise:waitfor_Promise<boolean>)=>Promise<number>)}[period]
  * @returns {Promise<boolean>}
  */
 export function file_exists__waitfor(
-	path,
+	path_OR_op,
 	timeout,
 	period
 ) {
-	return waitfor(()=>
-		file_exists_(path),
-	timeout ?? 5000,
-	period ?? 0)
+	throw new Error('no browser support')
 }