+++ all/file_exists/index.d.ts
@@ -4,7 +4,7 @@ export {
 	file_exists_ as path__exists_
 }
 export declare function file_exists__waitfor(
-	path:string,
+	path_OR_op:string|(()=>unknown|Promise<unknown>),
 	timeout?:number,
 	period?:number|((promise:waitfor_Promise<boolean>)=>Promise<number>)
 ):waitfor_Promise<boolean>